refactor getpublisheddatetime into shared module
